Refacter
classobj has 2 new lists for methods and attributes oeach class contains
Class for parsing files returns a ClassForCLass obj.
GlobalHandler has a new List which will store all the parsed classes.
Fix the display now it works as we expect
Need to fix where the Classes are in the display and to connect the
depends on their connection
Also need new photos and add the IFs for interface and abstract classes